Custom info box + marker and added from and to labels
Implemented a custom marker info window asking the user if the current
location is their desired pickup location. Also added the custom marker
icon and to and from label boxes. Still need to create a wrapper class
to allow the button clicks in the info window to be received, as buttons
in this window is not native.