Added in Search Mode and Map Mode
Implemented in a search mode, which instead of being an actual activity
is simply an interface blocking the map. Allows user to input address
and select from a list of results.

Map mode is the mode in which the user is previewing a location for
selection. It presents the selected location's map image and the
(direct) distance to the location. I'll need to figure out how to get
the distance by street soon.

Also added in some new button  design stuff.